Refactored section modules, added pets section
