+++ assets/js/main.js
@@ -9,14 +9,13 @@ import { HandlerFactory } from '{{ "assets/js/modules/handler.js" | relative_url
     const _handler = HandlerFactory();
 
     const _sections = {
-        [_handler.sections.achievements]: {
-            'json': '{{ "assets/json/achievements.json" | relative_url}}'
-        }
+        [_handler.sections.achievements]: '{{ "assets/json/achievements.json" | relative_url }}',
+        [_handler.sections.pets]: '{{ "assets/json/pets.json" | relative_url }}',
     };
 
     //
     for(const key of Object.keys(_sections)) {
-        $.getJSON(_sections[key].json, function(data) {
+        $.getJSON(_sections[key], function(data) {
             _handler.initSection(this, data);
         }.bind(key));
     }
