+++ assets/js/modules/factories.js
@@ -24,6 +24,8 @@ class Store {
         this.isUnlocked = this.isUnlocked.bind(this);
     }
 
+    //init
+    //Initialise @obj using localStorage
     init() {
         //Update @_userObj with localStorage if available
         if (localStorage.getItem(this.id) !== null) {
@@ -35,6 +37,9 @@ class Store {
         return this;
     }
 
+    //isUnlocked
+    //Check whether or not an item is unlocked based on @requirements
+    //@requirements - object containing item requirements
     isUnlocked(requirements) {
         if (requirements) {
             if ('combat' in requirements && (requirements.combat && !this.obj.complete.skills.includes('Combat'))) {
@@ -159,7 +164,60 @@ class Store {
     }
 }
 
-export function StoreFactory() {
+class Skills {
+    constructor(onClick) {
+        this.onClick = onClick;
+
+        this.selectors = {
+            items: '#skills-wrapper>div',
+            activeItems: '#skills-wrapper>div._active>._id:not(:contains("Combat"))',
+            progress: '#skills-progress'
+        };
+    }
+
+    //init
+    //Initialise skill items status (_active, _inactive) based on @completed and set click event handlers
+    //@complete - localStorage completed skills object
+    init(completed) {
+        $(this.selectors.items).each((index, element) => {
+            if (completed.includes($(element).find('._id').text())) {
+                $(element).addClass('_active');
+                $(element).removeClass('_inactive');
+            }
+        });
+
+        //Skills items click event handler
+        $(this.selectors.items).on('click', (event) => {
+
+            //Toggle _active and _inactive classes to visually update items
+            $(event.currentTarget).toggleClass('_active');
+            $(event.currentTarget).toggleClass('_inactive');
+
+            //Item click callback with skill name
+            this.onClick($(event.currentTarget).find('._id').text());
+            this.updateProgress();
+        });
+
+        this.updateProgress();
+        return this;
+    }
+
+
+    //updateProgress
+    //Update the skill progress count
+    updateProgress() {
+        $(this.selectors.progress).text($(this.selectors.activeItems).length);
+    }
+}
+
+function SkillsFactory(onClick, completed) {
+    const skills = new Skills(onClick);
+    return skills.init(completed);
+}
+
+function StoreFactory() {
     const store = new Store();
     return store.init();
 }
+
+export { SkillsFactory, StoreFactory }
